Finish converting to using pread64

+++ native-unpacker/kisskiss.h
@@ -26,6 +26,6 @@ uint32_t get_clone_pid(uint32_t service_pid);
 uint32_t get_process_pid(const char* target_package_name);
 char *determine_filter(uint32_t clone_pid, int memory_fd);
 int find_magic_memory(uint32_t clone_pid, int memory_fd, memory_region *memory, char* extra_filter);
-int peek_memory(int memory_file, uint32_t address);
+int peek_memory(int memory_file, uint64_t address);
 int dump_memory(int memory_fd, memory_region *memory, const char* file_name);
 int attach_get_memory(uint32_t pid);
